Remove backend provision. Error if not provided.
This library should only provide backend methods
that don't define a data structure.

This commit sets the backend to a noop, gives an
example for projects to provide their own backend,
and errors if required methods aren't provided.

It also allows test/*-mock.js files to be loaded
before the test/*-spec.js files so that common
code can be shared amongst the tests. The first
example of this is a mocking of the backend to
suppress errors that are now being thrown.